Apply consistent folder/namespace for tests of home screen display logic, to enable a class rename in the next commit.
