+++ Application/EdFi.Ods.AdminApp.Management.Tests/Controllers/Display/HomeScreen/WhenDisplayingOnPremisesHomeScreen.cs
@@ -8,33 +8,33 @@ using EdFi.Ods.AdminApp.Web.Display.DisplayService;
 using EdFi.Ods.AdminApp.Web.Display.HomeScreen;
 using NUnit.Framework;
 
-namespace EdFi.Ods.AdminApp.Management.Tests.Display.HomeScreenDisplayService
+namespace EdFi.Ods.AdminApp.Management.Tests.Controllers.Display.HomeScreen
 {
     [TestFixture]
-    public class WhenDisplayingAzureHomeScreen
+    public class WhenDisplayingOnPremisesHomeScreen
     {
         private static HomeScreenDisplayAssertions HomeScreenTabAssertions(params Permission[] permissions)
         {
-            return new HomeScreenDisplayAssertions(new AzureHomeScreenDisplayService(new AdminAppUserContext
+            return new HomeScreenDisplayAssertions(new OnPremHomeScreenDisplayService(new AdminAppUserContext
             {
                 Permissions = permissions
             }));
         }
 
         [Test]
-        public void Should_display_the_settings_button()
+        public void Should_display_settings_tab()
         {
             HomeScreenTabAssertions().TabShouldBeEnabled(HomeScreenEnumeration.Settings);
         }
 
         [Test]
-        public void Should_display_global_button()
+        public void Should_display_production_global_tab()
         {
             HomeScreenTabAssertions(Permission.AccessGlobalSettings).TabShouldBeEnabled(HomeScreenEnumeration.Global);
         }
 
         [Test]
-        public void Should_not_display_global_button()
+        public void Should_not_display_production_global_tab()
         {
             HomeScreenTabAssertions().TabShouldNotBeEnabled(HomeScreenEnumeration.Global);
         }
